+++ src/all/kavita/CHANGELOG.md
@@ -1,3 +1,14 @@
+## 1.3.12
+
+## Features
+
+* Migrate filters to v2
+* Implemented smartFilters
+* Added localization support
+
+## Fixed
+* Fixed publication status not showing
+
 ## 1.3.10
 
 ### Features